requirements.txt
flask==2.3.3
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Werkzeug>=2.3.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
safetensors>=0.3.0
tokenizers>=0.13.0
requests>=2.31.0
PyMuPDF
python-docx